Split implementation out of PlatformWebView on UWP.
