+++ src/Auth0.OidcClient.UWP/Auth0.OidcClient.UWP.csproj
@@ -110,6 +110,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Auth0Client.cs" />
+    <Compile Include="WebAuthenticationBrokerBrowser.cs" />
     <Compile Include="PlatformWebView.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="Properties\Auth0.OidcClient.UWP.rd.xml" />
